# specs/changes/active/20260912-iam-identity-access-management/contracts/internal-openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "IAM资源内部授权契约",
    "version": "2.0.0"
  },
  "paths": {
...
    "/inner/authorization/v2/evaluate": {
      "post": {
        "operationId": "iamEvaluateV2",
        "description": "INNER受信任调用及在线原始身份；请求不能替代用户；注册操作模式决定fresh或读热缓存",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthorizationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "R信封；精确错误码区分拒绝和不可用",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RAuthorizationDecision"
                }
              }
            }
          },
          "400": {
            "description": "稳定业务错误码；失败时data为空",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IamErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "稳定业务错误码；失败时data为空",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IamErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "稳定业务错误码；失败时data为空",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IamErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "稳定业务错误码；失败时data为空",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IamErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "稳定业务错误码；失败时data为空",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IamErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "稳定业务错误码；失败时data为空",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IamErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/inner/authorization/v2/preview": {
      "post": {
        "operationId": "iamPreviewV2",
        "description": "交互预览可使用短期读缓存；不能代替业务事务中的evaluate最终写校验",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthorizationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "R信封；精确错误码区分拒绝和不可用",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RAuthorizationDecision"
                }
              }
            }
          },
          "400": {
            "description": "稳定业务错误码；失败时data为空",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IamErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "稳定业务错误码；失败时data为空",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IamErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "稳定业务错误码；失败时data为空",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IamErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "稳定业务错误码；失败时data为空",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IamErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "稳定业务错误码；失败时data为空",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IamErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "稳定业务错误码；失败时data为空",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IamErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ActionDecision": {
        "required": [
          "fields"
        ],
        "type": "object",
        "properties": {
          "allowed": {
            "type": "boolean"
          },
          "governed": {
            "type": "boolean"
          },
          "scope": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ScopeCondition"
            }
          },
          "fields": {
            "$ref": "#/components/schemas/FieldPolicyDecision"
          }
        }
      },
...
      "AuthorizationContext": {
        "required": [
          "accountId",
          "domain",
          "memberId"
        ],
        "type": "object",
        "properties": {
          "domain": {
            "type": "string",
            "description": "当前管理域",
            "enum": [
              "PLATFORM",
              "TENANT"
            ]
          },
          "tenantId": {
            "type": "string",
            "description": "租户域必填；平台域为空"
          },
          "accountId": {
            "type": "string",
            "description": "全局账号 ID"
          },
          "memberId": {
            "type": "string",
            "description": "当前域的成员 ID"
          }
        },
        "description": "经认证选择的当前成员身份；不接受客户端直接指定权限域",
        "readOnly": true
      },
      "AuthorizationDecision": {
        "type": "object",
        "properties": {
          "resource": {
            "$ref": "#/components/schemas/ResourceKey"
          },
          "context": {
            "$ref": "#/components/schemas/AuthorizationContext"
          },
          "actions": {
            "type": "object",
            "additionalProperties": {
              "$ref": "#/components/schemas/ActionDecision"
            }
          },
          "version": {
            "type": "string"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "AuthorizationRequest": {
        "required": [
          "actionCodes",
          "resource"
        ],
        "type": "object",
        "properties": {
          "resource": {
            "$ref": "#/components/schemas/ResourceKey"
          },
          "actionCodes": {
            "maxItems": 100,
            "minItems": 0,
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "FieldAccess": {
        "required": [
          "visibility"
        ],
        "type": "object",
        "properties": {
          "visibility": {
            "type": "string",
            "description": "字段可见程度",
            "enum": [
              "HIDDEN",
              "MASKED",
              "FULL"
            ]
          },
          "editable": {
            "type": "boolean",
            "description": "是否允许编辑，HIDDEN 必须为 false"
          }
        },
        "description": "描述服务端计算后的字段输出与编辑能力，不独立授予对象或写操作权限"
      },
...
      "FieldOperations": {
        "type": "object",
        "properties": {
          "editable": {
            "type": "boolean"
          },
          "filterable": {
            "type": "boolean"
          }
        }
      },
      "FieldPolicyDecision": {
        "type": "object",
        "properties": {
          "defaults": {
            "type": "object",
            "additionalProperties": {
              "$ref": "#/components/schemas/FieldAccess"
            }
          },
          "ceilings": {
            "type": "object",
            "additionalProperties": {
              "$ref": "#/components/schemas/FieldAccess"
            }
          },
          "rules": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ResolvedFieldRule"
            }
          },
          "operations": {
            "type": "object",
            "additionalProperties": {
              "$ref": "#/components/schemas/FieldOperations"
            }
          },
          "masks": {
            "type": "object",
            "additionalProperties": {
              "$ref": "#/components/schemas/MaskSpec"
            }
          },
          "mergeMode": {
            "type": "string",
            "enum": [
              "RESTRICTIONS",
              "GRANTS"
            ]
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "IamErrorEnvelope": {
        "type": "object",
        "required": [
          "code",
          "message",
          "data",
          "success"
        ],
        "properties": {
          "code": {
            "type": "string",
            "enum": [
              "InvalidArgument",
              "IdentityInvalid",
              "ActionDenied",
              "PasswordChangeRequired",
              "DataScopeDenied",
              "DelegationExceeded",
              "RoleRevisionUnavailable",
              "ApplicationUnavailable",
              "ObjectNotFound",
              "PolicyConflict",
              "RevisionConflict",
              "ObjectInUse",
              "StepUpFailed",
              "AuthorizationUnavailable"
            ]
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "object",
            "nullable": true,
            "enum": [
              null
            ]
          },
          "success": {
            "type": "boolean",
            "enum": [
              false
            ]
          }
        }
      },
      "MaskSpec": {
        "required": [
          "kind"
        ],
        "type": "object",
        "properties": {
          "kind": {
            "type": "string",
            "enum": [
              "PHONE",
              "EMAIL",
              "ALL",
              "KEEP_EDGES",
              "RANGE"
            ]
          },
          "prefix": {
            "type": "integer",
            "format": "int32"
          },
          "suffix": {
            "type": "integer",
            "format": "int32"
          },
          "start": {
            "type": "integer",
            "format": "int32"
          },
          "end": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "RAuthorizationDecision": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/AuthorizationDecision"
          },
          "success": {
            "type": "boolean"
          }
        }
      },
...
      "ResolvedFieldRule": {
        "type": "object",
        "properties": {
          "fieldKey": {
            "type": "string"
          },
          "scope": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ScopeCondition"
            }
          },
          "access": {
            "$ref": "#/components/schemas/FieldAccess"
          }
        }
      },
...
      "ResourceKey": {
        "required": [
          "applicationCode",
          "domain",
          "resourceCode"
        ],
        "type": "object",
        "properties": {
          "domain": {
            "type": "string",
            "enum": [
              "PLATFORM",
              "TENANT"
            ]
          },
          "applicationCode": {
            "type": "string"
          },
          "resourceCode": {
            "type": "string"
          }
        }
      },
...
      "ScopeCondition": {
        "type": "object",
        "properties": {
          "all": {
            "type": "boolean"
          },
          "objectIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "ownerMemberId": {
            "type": "string"
          },
          "departmentSets": {
            "type": "array",
            "items": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}